# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "srvcs-vectoradd",
    "description": "",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/": {
      "get": {
        "tags": [],
        "summary": "`GET /` — service identity (srvcs service standard).",
        "operationId": "index",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Info"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [],
        "summary": "`POST /` — component-wise vector addition.",
        "description": "This service owns the *control flow* but delegates every arithmetic step to\n`srvcs-floatadd`. The two operand vectors must have equal length (else\n`422`). For each index `i` it computes `floatadd(a[i], b[i]).result` and\npushes it onto the result list.\n\nElement-level validation is not handled here: this service never calls\n`srvcs-isnumber` directly. If `floatadd` is unreachable it reports itself\ndegraded (`503`); if `floatadd` rejects a component it forwards the `422`.",
        "operationId": "evaluate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VectorAddResponse"
                }
              }
            }
          },
          "422": {
            "description": "vector length mismatch, or a dependency rejected a component (forwarded)"
          },
          "500": {
            "description": "a dependency returned a malformed result"
          },
          "503": {
            "description": "a dependency is unavailable"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "EvalRequest": {
        "type": "object",
        "required": [
          "a",
          "b"
        ],
        "properties": {
          "a": {
            "type": "object",
            "description": "The first vector: a JSON array of numbers."
          },
          "b": {
            "type": "object",
            "description": "The second vector: a JSON array of numbers. Must match `a` in length."
          }
        }
      },
      "Info": {
        "type": "object",
        "required": [
          "service",
          "concern",
          "depends_on"
        ],
        "properties": {
          "concern": {
            "type": "string"
          },
          "depends_on": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "service": {
            "type": "string"
          }
        }
      },
      "VectorAddResponse": {
        "type": "object",
        "required": [
          "a",
          "b",
          "result"
        ],
        "properties": {
          "a": {
            "type": "object"
          },
          "b": {
            "type": "object"
          },
          "result": {
            "type": "object",
            "description": "The component-wise sum, a JSON array of `f64` numbers."
          }
        }
      }
    }
  }
}